Allows external control for Rx Complete loop in sync mode
The STM32 subGHz interrupt handler is already splitting the different IRQ event and clearing the event register. Using the SX126X_RF_API_process function is complex in this situation. But it is simple to rewrite it.
As a consequence in synchronous mode, the IRQ handler will work the same way as in async mode and the sync loop here must be able to exit when the Rx has been completed. This will by the  switch of sx126x_cts.rx_done_flag asynchronously.
Functionnaly speaking, this modification does not change anything in the current code (rx_done_flag set by SX126X_RF_API_process, so sequentially in synchronous mode) but allows the approach described above.