Allows to overide sx126x_get_and_clear_irq_status w/o impacting semtech lib
Use of this function is conflicting when using the subghz interrupt handler already defined in stm32wl HAL. For this implementation the process function can't get the values from the register already cleared. The function can be overide to match the process expectation. This is a way to limit the impact on the code.